requirements.txt
astroid
black
chromadb
faiss-cpu
google-generativeai
langchain
langchain-community
langgraph
mypy
openai
pylint
pymysql
pytest
python-dotenv
streamlit
pygments
tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-html
tree-sitter-css
python-magic
chardet